+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/store/goods/StoreGoodsListQueryParam.java
@@ -15,7 +15,7 @@ import lombok.Data;
 @Data
 public class StoreGoodsListQueryParam {
 	
-	private Short catId;
+	private Integer catId;
 	private Byte isOnSale;
 	
 	/**
